// frontend/src/types.ts
export interface StartupItem {
    id: string;
    name: string;
    logoUrl: string;
    stage: string;
    status: string;
    member?: string;
    tagline: string;
    statement: string;
    description: string;
    usp: string;
    openPositions: string[];
    aboutUs: string;
    team: string;
    jobDescription: string;
  }
  